+++ plugins/collide/opcode/OPC_TreeCollider.cpp
@@ -274,7 +274,7 @@ void AABBTreeCollider::InitQuery(const Matrix4x4* world0, const Matrix4x4* world
 		for(udword j=0;j<3;j++)
 		{
 			// Epsilon value prevents floating-point inaccuracies (strategy borrowed from RAPID)
-			mAR.m[i][j] = 1e-6f + fabs(mR1to0.m[i][j]);
+			mAR.m[i][j] = 1e-6f + (float)fabs(mR1to0.m[i][j]);
 		}
 	}
 }
